renamed the nutrition tracker module from userNutritionTracker to nutritionTracker added the ExerciseTrackerModule and NutritionTrackerModule in the app.module imports.
