+++ the-pack-backend/src/app.module.ts
@@ -11,7 +11,7 @@ import { ClassModule } from './classes/class.module';
 import { userClassesModule } from './userClasses/userClasses.module';
 import { schedulingModule } from './scheduling/scheduling.module';
 import { ExerciseTrackerModule } from './userExerciseTracker/userExerciseTracker.module';
-import { NutritionTrackerModule } from './userNutritionTracker/userNutritionTracker.module';
+import { NutritionTrackerModule } from './nutritionTracker/nutritionTracker.module';
 import { ConfigModule } from '@nestjs/config';
 import { ChatroomModule } from './chatroom/chatroom.module';
 import { MessagesModule } from './messages/messages.module';
@@ -20,7 +20,7 @@ import { MessagesModule } from './messages/messages.module';
   
   imports: [ConfigModule.forRoot({isGlobal: true}),PrismaModule, UserModule, CoachModule, 
     AnnouncementsModule, ClassModule, userClassesModule, schedulingModule, 
-    AuthModule, ChatroomModule, MessagesModule],
+    AuthModule, ChatroomModule, MessagesModule, NutritionTrackerModule, ExerciseTrackerModule],
   controllers: [AppController],
   providers: [AppService, PrismaService],
 })
