+++ the-pack-backend/src/nutritionTracker/nutritionTracker.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { NutritionTrackerService } from './userNutritionTracker.service';
-import { NutritionTrackerController } from './userNutritionTracker.controller';
+import { NutritionTrackerService } from './nutritionTracker.service';
+import { NutritionTrackerController } from './nutritionTracker.controller';
 @Module({
 
     imports: [PrismaModule],
